refactor: relocate API schemas
