+++ python-service/app/models/__init__.py
@@ -1,5 +1,1 @@
-"""Models package initialization."""
-
-from .jobspy import ScrapedJob  # noqa: F401
-from .responses import StandardResponse, create_success_response, create_error_response  # noqa: F401
-from .evaluations import PersonaEvaluation, Decision, EvaluationSummary  # noqa: F401
+"""ORM models package."""
