+++ python-service/app/services/ai/gemini.py
@@ -8,7 +8,7 @@ import asyncio
 from loguru import logger
 
 from ...core.config import get_settings
-from ...models.responses import StandardResponse, create_success_response, create_error_response
+from ...schemas.responses import StandardResponse, create_success_response, create_error_response
 
 
 class GeminiService:
