+++ python-service/app/services/infrastructure/job_persistence.py
@@ -10,7 +10,7 @@ import json
 from loguru import logger
 
 from ...core.config import get_settings
-from ...models.jobspy import ScrapedJob
+from ...schemas.jobspy import ScrapedJob
 from .database import get_database_service
