+++ python-service/app/services/infrastructure/postgrest.py
@@ -7,7 +7,7 @@ import httpx
 from loguru import logger
 
 from ...core.config import get_settings
-from ...models.responses import StandardResponse, create_success_response, create_error_response
+from ...schemas.responses import StandardResponse, create_success_response, create_error_response
 
 
 class PostgRESTService:
