+++ python-service/test_job_persistence.py
@@ -12,7 +12,7 @@ from datetime import datetime, timezone
 # Add the project root to Python path
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from app.models.jobspy import ScrapedJob
+from app.schemas.jobspy import ScrapedJob
 from app.services.infrastructure.job_persistence import JobPersistenceService, persist_jobs
