+++ python-service/test_job_persistence_integration.py
@@ -13,7 +13,7 @@ async def test_persist_jobs_inserts_row():
         os.environ["DATABASE_URL"] = pg.url()
 
         # Import after setting DATABASE_URL so services pick it up
-        from app.models.jobspy import ScrapedJob
+        from app.schemas.jobspy import ScrapedJob
         from app.services.infrastructure.job_persistence import persist_jobs
 
         # Create jobs table schema
